fw/services/activity: widen ActivityScalarStore to uint32_t
Daily step count and distance metrics were stored as uint16_t, so the
watch UI wrapped at 65,535 steps (e.g. 79,396 displayed as 13,860) and
at 65.5 km of daily distance. The kraepelin algorithm already tracks
steps in an int32_t; the value was truncated when copied into the
metric store. The phone app was unaffected because it sums raw minute
records itself.

Widen ActivityScalarStore to uint32_t and migrate the activity
settings file (version 3): history records and scalar metric records
are expanded from uint16_t to uint32_t values on first boot after the
update; all other records are copied verbatim. Settings files of
versions 1 and 2 both migrate through the same rewrite pass.

Fixes FIRM-3071

+++ include/pbl/services/activity/activity_algorithm.h
@@ -165,7 +165,7 @@ void activity_algorithm_minute_handler(time_t utc_sec, AlgMinuteRecord *record_o
 //! Return the current number of steps computed
 //! @param[out] steps the number of steps is returned in this variable
 //! @return true if success
-bool activity_algorithm_get_steps(uint16_t *steps);
+bool activity_algorithm_get_steps(uint32_t *steps);
 
 //! Tells the activity algorithm whether or not it should automatically track activities
 //! @param enable true to start tracking, false to stop tracking
@@ -190,7 +190,7 @@ bool activity_algorithm_metrics_changed_notification(void);
 //! a watch reboot.
 //! @param[in] steps set the number of steps to this
 //! @return true if success
-bool activity_algorithm_set_steps(uint16_t steps);
+bool activity_algorithm_set_steps(uint32_t steps);
 
 //! Return the timestamp of the last minute that was processed by the sleep detector.
 time_t activity_algorithm_get_last_sleep_utc(void);
